[Refactor] Move rpc files to its own folder
