+++ src/rpc/blockchain.cpp
@@ -10,7 +10,7 @@
 #include "checkpoints.h"
 #include "clientversion.h"
 #include "main.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "sync.h"
 #include "txdb.h"
 #include "util.h"
