+++ src/rpc/fundamentalnode.cpp
